pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "razorpay>=1.4.2",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "trafilatura>=2.0.0",
    "uvicorn>=0.35.0",
]
